+++ tests/test_flag/test_flag_short.py
@@ -9,17 +9,17 @@ from pydantic import BaseModel
 class FlagShort(BaseModel):
     """Model with a flag with shorthand"""
 
-    proceed: Annotated[bool, short("a")]
+    proceed: Annotated[bool, short("p")]
     """Continue with this operation."""
 
 
-@pytest.mark.parametrize("args", (["--proceed"], ["-a"]))
+@pytest.mark.parametrize("args", (["--proceed"], ["-p"]))
 def test_short_hand_flag_true(args):
     model = parse(FlagShort, args)
     assert model == FlagShort(proceed=True)
 
 
-@pytest.mark.parametrize("args", (["--no-proceed"], ["-no-a"]))
+@pytest.mark.parametrize("args", (["--no-proceed"], ["-no-p"]))
 def test_short_hand_flag_false(args):
     model = parse(FlagShort, args)
     assert model == FlagShort(proceed=False)
@@ -31,6 +31,13 @@ def test_help(capture_output):
     assert err.value.code == 0
 
     assert (
-        "-no-a/-a/--proceed/--no-proceed Continue with this operation. [bool]"
-        in capture_output.captured_output
+        """
+Usage: my-cli-app [Arguments]
+
+Model with a flag with shorthand
+
+Arguments:
+    -no-p/-p --proceed/--no-proceed Continue with this operation. [bool]
+"""
+        == capture_output.captured_output
     )